[refactor] Split undoable commands into separate source files
